helpers.foldline: do not insert spaces between the high and low UTF-16 surrogates
ical.js used to fold in the middle of a UTF-16 code point sequence.  To be
precise, it inserted a space and a new line after the high UTF-16 surrogate
and before the low UTF-16 surrogate.  This created invalid stand-alone
surrogates which the JavaScript engine converted to something valid

This is the problem, with Node.js:

import {writeFileSync} from 'fs';
const a = '\uD83D\uDCAA'  //this is 💪'Flexed Biceps', UTF-8 encoding: 0xF0 0x9F
// 0x92 0xAA, HTML Entity: &#128170; &#x1F4AA; UTF-16 encoding: 0xD83D 0xDCAA
writeFileSync('a1', a, 'utf8')  //file contains the bytes f0 9f 92 aa
writeFileSync('a2', a.charAt(0) + a.charAt(1), 'utf8')  //file contains the bytes f0 9f 92 aa
writeFileSync('b', a.charAt(0) + ' ' + a.charAt(1), 'utf8') //file contains the bytes
0xEF 0xBF 0xBD 0x20 0xEF 0xBF 0xBD = REPLACEMENT CHARACTER = �

so the result in b is valid UTF-8, but it has nothing in common with the
original text and it cannot be suspected, that anyform of reconstructing b
to a shall be tried.

This patch takes a full UTF-16 character from the input,
• calculates whether it takes one or two UTF-16 chars, and keeps in pos
where the next full UTF-16 character starts,
• calculates for the full UTF-16 code point, whether it needs 1, 2, 3 or 4
bytes to be presened in UTF-8, keeping in line_length the bytes for UTF-8
so far necessary,
• splits the line, when the UTF-8 presentation exceeds ICAL.foldLength bytes